v1.0.0

+++ scripts/resolve_frontend.sh
@@ -53,17 +53,20 @@ filename=$(basename $url)
 echo "final url: $url"
 echo "filename: $filename"
 
-# 下载zip文件
-wget $url
+# 如果第一个参数为true，则下载zip文件
+if [ "$1" = "true" ]; then
+  # 下载zip文件
+  wget $url
 
-# 解压zip文件
-unzip -o $filename
+  # 解压zip文件
+  unzip -o $filename
 
-# 删除下载的zip文件
-rm $filename
+  # 删除下载的zip文件
+  rm $filename
 
-# 移动到指定目录
-mv dist/* labelu/internal/statics
+  # 移动到指定目录
+  mv dist/* labelu/internal/statics
 
-# 删除空目录
-rm -rf dist
+  # 删除空目录
+  rm -rf dist
+fi
